Dev Branch Created
To save my current work. Things are... VERY broken right now. Will probably revert back to what's on the main branch  and restart from there. This is unsalvageable without more work than it's worth, but I do want to keep it for posterity as there is some useful code in here.